Break devices.py into multiple files (#3064)
What is the motivation for this PR?
The devices.py script is getting too big over time.

How did you do it?
This change is to break it into multiple smaller modules for better readability and manageability.

How did you verify/test it?
Tested several scripts on KVM. The change should can be covered better by PR test.
