+++ tests/common/devices/eos.py
@@ -1,0 +1,171 @@
+import ipaddress
+import json
+import logging
+
+from tests.common.devices.base import AnsibleHostBase
+
+logger = logging.getLogger(__name__)
+
+
+class EosHost(AnsibleHostBase):
+    """
+    @summary: Class for Eos switch
+
+    For running ansible module on the Eos switch
+    """
+
+    def __init__(self, ansible_adhoc, hostname, eos_user, eos_passwd, shell_user=None, shell_passwd=None, gather_facts=False):
+        '''Initialize an object for interacting with EoS type device using ansible modules
+
+        Args:
+            ansible_adhoc (): The pytest-ansible fixture
+            hostname (string): hostname of the EOS device
+            eos_user (string): Username for accessing the EOS CLI interface
+            eos_passwd (string): Password for the eos_user
+            shell_user (string, optional): Username for accessing the Linux shell CLI interface. Defaults to None.
+            shell_passwd (string, optional): Password for the shell_user. Defaults to None.
+            gather_facts (bool, optional): Whether to gather some basic facts. Defaults to False.
+        '''
+        self.eos_user = eos_user
+        self.eos_passwd = eos_passwd
+        self.shell_user = shell_user
+        self.shell_passwd = shell_passwd
+        AnsibleHostBase.__init__(self, ansible_adhoc, hostname)
+        self.localhost = ansible_adhoc(inventory='localhost', connection='local', host_pattern="localhost")["localhost"]
+
+    def __getattr__(self, module_name):
+        if module_name.startswith('eos_'):
+            evars = {
+                'ansible_connection':'network_cli',
+                'ansible_network_os':'eos',
+                'ansible_user': self.eos_user,
+                'ansible_password': self.eos_passwd,
+                'ansible_ssh_user': self.eos_user,
+                'ansible_ssh_pass': self.eos_passwd,
+                'ansible_become_method': 'enable'
+            }
+        else:
+            if not self.shell_user or not self.shell_passwd:
+                raise Exception("Please specify shell_user and shell_passwd for {}".format(self.hostname))
+            evars = {
+                'ansible_connection':'ssh',
+                'ansible_network_os':'linux',
+                'ansible_user': self.shell_user,
+                'ansible_password': self.shell_passwd,
+                'ansible_ssh_user': self.shell_user,
+                'ansible_ssh_pass': self.shell_passwd,
+                'ansible_become_method': 'sudo'
+            }
+        self.host.options['variable_manager'].extra_vars.update(evars)
+        return super(EosHost, self).__getattr__(module_name)
+
+    def shutdown(self, interface_name):
+        out = self.eos_config(
+            lines=['shutdown'],
+            parents='interface %s' % interface_name)
+        logging.info('Shut interface [%s]' % interface_name)
+        return out
+
+    def no_shutdown(self, interface_name):
+        out = self.eos_config(
+            lines=['no shutdown'],
+            parents='interface %s' % interface_name)
+        logging.info('No shut interface [%s]' % interface_name)
+        return out
+
+    def check_intf_link_state(self, interface_name):
+        show_int_result = self.eos_command(
+            commands=['show interface %s' % interface_name])
+        return 'Up' in show_int_result['stdout_lines'][0]
+
+    def set_interface_lacp_rate_mode(self, interface_name, mode):
+        out = self.eos_config(
+            lines=['lacp rate %s' % mode],
+            parents='interface %s' % interface_name)
+
+        if out['failed'] == True:
+            # new eos deprecate lacp rate and use lacp timer command
+            out = self.eos_config(
+                lines=['lacp timer %s' % mode],
+                parents='interface %s' % interface_name)
+            if out['changed'] == False:
+                logging.warning("Unable to set interface [%s] lacp timer to [%s]" % (interface_name, mode))
+                raise Exception("Unable to set interface [%s] lacp timer to [%s]" % (interface_name, mode))
+            else:
+                logging.info("Set interface [%s] lacp timer to [%s]" % (interface_name, mode))
+        else:
+            logging.info("Set interface [%s] lacp rate to [%s]" % (interface_name, mode))
+        return out
+
+    def kill_bgpd(self):
+        out = self.eos_config(lines=['agent Rib shutdown'])
+        return out
+
+    def start_bgpd(self):
+        out = self.eos_config(lines=['no agent Rib shutdown'])
+        return out
+
+    def check_bgp_session_state(self, neigh_ips, neigh_desc, state="established"):
+        """
+        @summary: check if current bgp session equals to the target state
+
+        @param neigh_ips: bgp neighbor IPs
+        @param neigh_desc: bgp neighbor description
+        @param state: target state
+        """
+        neigh_ips = [ip.lower() for ip in neigh_ips]
+        neigh_ips_ok = []
+        neigh_desc_ok = []
+        neigh_desc_available = False
+
+        out_v4 = self.eos_command(
+            commands=['show ip bgp summary | json'])
+        logging.info("ip bgp summary: {}".format(out_v4))
+
+        out_v6 = self.eos_command(
+            commands=['show ipv6 bgp summary | json'])
+        logging.info("ipv6 bgp summary: {}".format(out_v6))
+
+        for k, v in out_v4['stdout'][0]['vrfs']['default']['peers'].items():
+            if v['peerState'].lower() == state.lower():
+                if k in neigh_ips:
+                    neigh_ips_ok.append(k)
+                if 'description' in v:
+                    neigh_desc_available = True
+                    if v['description'] in neigh_desc:
+                        neigh_desc_ok.append(v['description'])
+
+        for k, v in out_v6['stdout'][0]['vrfs']['default']['peers'].items():
+            if v['peerState'].lower() == state.lower():
+                if k.lower() in neigh_ips:
+                    neigh_ips_ok.append(k)
+                if 'description' in v:
+                    neigh_desc_available = True
+                    if v['description'] in neigh_desc:
+                        neigh_desc_ok.append(v['description'])
+        logging.info("neigh_ips_ok={} neigh_desc_available={} neigh_desc_ok={}"\
+            .format(str(neigh_ips_ok), str(neigh_desc_available), str(neigh_desc_ok)))
+        if neigh_desc_available:
+            if len(neigh_ips) == len(neigh_ips_ok) and len(neigh_desc) == len(neigh_desc_ok):
+                return True
+        else:
+            if len(neigh_ips) == len(neigh_ips_ok):
+                return True
+
+        return False
+
+    def exec_template(self, ansible_root, ansible_playbook, inventory, **kwargs):
+        playbook_template = 'cd {ansible_path}; ansible-playbook {playbook} -i {inventory} -l {fanout_host} --extra-vars \'{extra_vars}\' -vvvvv'
+        cli_cmd = playbook_template.format(ansible_path=ansible_root, playbook=ansible_playbook, inventory=inventory,
+            fanout_host=self.hostname, extra_vars=json.dumps(kwargs))
+        res = self.localhost.shell(cli_cmd)
+
+        if res["localhost"]["rc"] != 0:
+            raise Exception("Unable to execute template\n{}".format(res["stdout"]))
+
+    def get_route(self, prefix):
+        cmd = 'show ip bgp' if ipaddress.ip_network(unicode(prefix)).version == 4 else 'show ipv6 bgp'
+        return self.eos_command(commands=[{
+            'command': '{} {}'.format(cmd, prefix),
+            'output': 'json'
+        }])['stdout'][0]
