+++ tests/common/devices/ptf.py
@@ -1,0 +1,13 @@
+from tests.common.devices.base import AnsibleHostBase
+
+
+class PTFHost(AnsibleHostBase):
+    """
+    @summary: Class for PTF
+
+    Instance of this class can run ansible modules on the PTF host.
+    """
+    def __init__(self, ansible_adhoc, hostname):
+        AnsibleHostBase.__init__(self, ansible_adhoc, hostname)
+
+    # TODO: Add a method for running PTF script
